+++ ukr/taggers/tokenize_and_classify.py
@@ -23,6 +23,7 @@ from ukr.taggers.score import ScoreFst
 from ukr.taggers.telephone import TelephoneFst
 from ukr.taggers.time import TimeFst
 from ukr.taggers.version import VersionFst
+from ukr.taggers.whitelist import WhitelistFst
 from ukr.taggers.word import WordFst
 
 
@@ -59,9 +60,11 @@ class ClassifyFst(GraphFst):
         ip_graph = IpFst(cardinal=cardinal).fst
         word_graph = WordFst().fst
         money_graph = MoneyFst(cardinal=cardinal, decimal=decimal).fst
+        whitelist_graph = WhitelistFst().fst
 
         classify = (
-                pynutil.add_weight(duration_graph, 1.09)
+                pynutil.add_weight(whitelist_graph, 1.01)
+                | pynutil.add_weight(duration_graph, 1.09)
                 | pynutil.add_weight(decade_graph, 1.09)
                 | pynutil.add_weight(legal_graph, 1.09)
                 | pynutil.add_weight(score_graph, 1.09)