+++ ukr/verbalizers/time.py
@@ -1,7 +1,7 @@
 import pynini
 from pynini.lib import pynutil
 
-from ukr.graph_utils import NEMO_DIGIT, GraphFst, delete_space
+from ukr.graph_utils import NEMO_CHAR, NEMO_DIGIT, GraphFst, delete_space, insert_space
 
 
 class TimeFst(GraphFst):
@@ -34,5 +34,15 @@ class TimeFst(GraphFst):
                 + pynutil.insert(":")
                 + (minute @ add_leading_zero_to_double_digit)
         )
+        zone = (
+                delete_space
+                + insert_space
+                + pynutil.delete("zone:")
+                + delete_space
+                + pynutil.delete('"')
+                + pynini.closure(NEMO_CHAR - " ", 1)
+                + pynutil.delete('"')
+        )
+        graph += pynini.closure(zone, 0, 1)
         delete_tokens = self.delete_tokens(graph)
         self.fst = delete_tokens.optimize()